Use ITSTrackingInterface in its-reco-wf
