+++ Detectors/ITSMFT/ITS/tracking/CMakeLists.txt
@@ -36,6 +36,15 @@ o2_add_library(ITStracking
                        O2::ITSMFTReconstruction
                        O2::DataFormatsITS)
 
+
+o2_add_library(ITSTrackingInterface
+               TARGETVARNAME targetName
+               SOURCES src/TrackingInterface.cxx
+               PUBLIC_LINK_LIBRARIES
+                       O2::ITStracking
+                       O2::Framework
+                       O2::GPUTracking)
+
 if (OpenMP_CXX_FOUND)
         target_compile_definitions(${targetName} PRIVATE WITH_OPENMP)
         target_link_libraries(${targetName} PRIVATE OpenMP::OpenMP_CXX)
