+++ Detectors/ITSMFT/ITS/tracking/include/ITStracking/Configuration.h
@@ -31,6 +31,12 @@ namespace o2
 namespace its
 {
 
+enum class TrackingMode {
+  Sync,
+  Async,
+  Cosmics
+};
+
 template <typename Param>
 class Configuration : public Param
 {
@@ -123,19 +129,11 @@ struct VertexingParameters {
 
 struct TimeFrameGPUParameters {
   TimeFrameGPUParameters() = default;
-  // TimeFrameGPUParameters(size_t cubBufferSize,
-  //                        size_t maxTrkClu,
-  //                        size_t cluLayCap,
-  //                        size_t cluROfCap,
-  //                        size_t maxTrkCap,
-  //                        size_t maxVertCap,
-  //                        size_t maxROFs);
 
   size_t tmpCUBBufferSize = 1e5; // In average in pp events there are required 4096 bytes
   size_t maxTrackletsPerCluster = 1e2;
   size_t clustersPerLayerCapacity = 2.5e5;
   size_t clustersPerROfCapacity = 1.5e3;
-  // size_t trackletsCapacity = maxTrackletsPerCluster * clustersPerROfCapacity;
   size_t validatedTrackletsCapacity = 1e3;
   size_t cellsLUTsize = validatedTrackletsCapacity;
   size_t maxNeighboursSize = 1e2;
@@ -147,29 +145,6 @@ struct TimeFrameGPUParameters {
   size_t nTimeFrameChunks = 3;
   int maxGPUMemoryGB = -1;
 };
-
-// inline TimeFrameGPUParameters::TimeFrameGPUParameters(size_t cubBufferSize,
-//                                                       size_t maxTrkClu,
-//                                                       size_t cluLayCap,
-//                                                       size_t cluROfCap,
-//                                                       size_t maxTrkCap,
-//                                                       size_t maxVertCap,
-//                                                       size_t maxROFs,
-//                                                       size_t validatedTrackletsCapacity,
-//                                                       size_t cellsLUTsize,
-//                                                       size_t maxNeighboursSize,
-//                                                       size_t neighboursLUTsize,
-//                                                       size_t maxRoadPerRofSize,
-//                                                       size_t maxLinesCapacity) : tmpCUBBufferSize{cubBufferSize},
-//                                                                                  maxTrackletsPerCluster{maxTrkClu},
-//                                                                                  clustersPerLayerCapacity{cluLayCap},
-//                                                                                  clustersPerROfCapacity{cluROfCap},
-//                                                                                  maxLinesCapacity{maxTrkCap},
-//                                                                                  maxVerticesCapacity{maxVertCap},
-//                                                                                  nMaxROFs{maxROFs}
-// {
-//   trackletsCapacity = maxTrackletsPerCluster * clustersPerLayerCapacity;
-// }
 
 } // namespace its
 } // namespace o2
