+++ Detectors/ITSMFT/ITS/workflow/include/ITSWorkflow/RecoWorkflow.h
@@ -16,9 +16,10 @@
 
 #include "Framework/WorkflowSpec.h"
 
-#include "GPUO2Interface.h"
-#include "GPUReconstruction.h"
-#include "GPUChainITS.h"
+// #include "GPUO2Interface.h"
+// #include "GPUReconstruction.h"
+// #include "GPUChainITS.h"
+#include "GPUDataTypes.h"
 
 namespace o2
 {
