+++ Detectors/ITSMFT/ITS/workflow/src/RecoWorkflow.cxx
@@ -31,6 +31,7 @@ namespace o2
 {
 namespace its
 {
+// using ITSTrackReconstruction::TrackingMode;
 
 namespace reco_workflow
 {
@@ -64,7 +65,7 @@ framework::WorkflowSpec getWorkflow(bool useMC,
         cfg.runITSTracking = true;
         cfg.itsTriggerType = useTrig;
         cfg.itsOverrBeamEst = overrideBeamPosition;
-        cfg.itsTrackingMode = trmode == "sync" ? 0 : (trmode == "async" ? 1 : 2);
+        cfg.itsTrackingMode = trmode == "sync" ? (int)TrackingMode::Sync : (trmode == "async" ? (int)TrackingMode::Async : (int)TrackingMode::Cosmics);
 
         Inputs ggInputs;
         auto ggRequest = std::make_shared<o2::base::GRPGeomRequest>(false, true, false, true, true,
